% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmt-ext.R
\name{add_lmt}
\alias{add_lmt}
\title{Install the Literate Markdown Tangler extension}
\usage{
add_lmt()
}
\value{
Nothing. Used for its side effects.
}
\description{
It installs the Quarto extension Literate Markdown Tangler, which provides
users with literate programming tools.
}
\examples{
\dontrun{
add_lmt()
}

}
