% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_objects.R
\name{battery_icon}
\alias{battery_icon}
\title{Generate a battery icon depending on input}
\usage{
battery_icon(x, max = 1000)
}
\arguments{
\item{max}{}
}
\description{
Generate a battery icon depending on input
}
