% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_objects.R
\name{latest_boxes}
\alias{latest_boxes}
\title{Title}
\usage{
latest_boxes(channel, formatting = NULL, icons = NULL,
  colors = "aqua", widths = 4, hrefs = NULL)
}
\arguments{
\item{channel}{A thingspeak channel}

\item{formatting}{}
}
\description{
Title
}
