% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{lmt}
\alias{lmt}
\title{Literate markdown tangler.}
\usage{
lmt(file, cache = FALSE)
}
\arguments{
\item{cache}{Boolean (\code{FALSE})}

\item{...}{List of file names.}
}
\description{
It tangles the code from a literate markdown file. If \code{cache} is set to
    \code{TRUE}, the function does nothing (useful if tangling is not needed
    every time the file is knitted.)
}
