% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{lmt_path}
\alias{lmt_path}
\title{Get system path to lmt.}
\usage{
lmt_path()
}
\value{
The path to the lmt executable.
}
\description{
The function returns the path to the lmt executable depending on the
    system.
}
\keyword{internal}
