% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praat.R
\name{praat_open}
\alias{praat_open}
\title{Open files with Praat.}
\usage{
praat_open(...)
}
\arguments{
\item{...}{A list of strings with the file paths.}
}
\description{
It opens a file or list of files in the Praat GUI.
}
\examples{
\dontrun{
# Open a single file
praat_open("./Desktop/your-script.praat")

# Open multiple diles
praat_open("./Desktop/script-1.praat", "./Desktop/script-2.praat")
}

}
