% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{praat_plot}
\alias{praat_plot}
\title{Plot waveform and spectrogram}
\usage{
praat_plot(file, wav, tg = NULL, start = 0, end = NULL, width = 5, f0 = FALSE)
}
\arguments{
\item{file}{Output file path as a string.}

\item{wav}{Wav file path as a string.}

\item{tg}{TextGrid file path as a string. If `NULL` (the default), a TextGrid
is plotted if a TextGrid file with the same name as the wav file is found.
No TextGrid is plotted otherwise.}

\item{start}{Start time of the plotting window in seconds.}

\item{end}{End time of the plotting window in seconds. If `NULL` (the
default), plot the entire duration.}

\item{width}{Width of the plot in inches.}

\item{f0}{Whether to plot f0 (`FALSE` by default).}
}
\value{
Nothing. It is used for its side effects.
}
\description{
Plot waveform and spectrogram
}
