% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{praatRun}
\alias{praatRun}
\title{Run a Praat script.}
\usage{
praatRun(script, ...)
}
\arguments{
\item{script}{Name (or path) of the script.}

\item{...}{List of arguments.}
}
\description{
It runs a file containing Praat scripting, with optional arguments.
}

