% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say.R
\name{say}
\alias{say}
\title{Tell R to say something !}
\usage{
say(message, voice = NULL)
}
\arguments{
\item{message}{A character string.}

\item{voice}{Voice name.}
}
\value{
Does not return anything. The function is only used for its side effect.
}
\description{
Give the function a message and R will pronounce it. Do not forget to turn on
your speakers!
}
\details{
R cannot speak by itself. The function uses speech synthesis implemented in
modern web browsers. In Rstudio, it uses the viewer panel. In other settings,
it attemps to open the default web browser.
}
\examples{
say("Hello world")

}
