% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say.R
\name{sayWidget}
\alias{sayWidget}
\title{a "HTMLwidget" that says something}
\usage{
sayWidget(message, voice = NULL, elementId = NULL)
}
\arguments{
\item{message}{A character string.}

\item{voice}{Voice name.}

\item{elementId}{Id of the HTML element that contains the widget.}
}
\description{
a "HTMLwidget" that says something
}
