% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{set_praat_path}
\alias{set_praat_path}
\title{Set path to Praat.}
\usage{
set_praat_path(path, default.path = FALSE)
}
\arguments{
\item{path}{The path to the Praat executable.}

\item{default.path}{If \code{TRUE}, the path is set to what \code{praatPath}
returns.}
}
\value{
Nothing. It is used for its side effects.
}
\description{
It assigns the default path to Praat or the user assigned path (when
    \code{default.path = FALSE}, which is the default).
}
