% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praatPath.R
\name{setPraatPath}
\alias{setPraatPath}
\title{Set path to Praat}
\usage{
setPraatPath(path, default.path = FALSE)
}
\arguments{
\item{path}{Path to Praat.}

\item{default.path}{Boolean (default = FALSE).}
}
\description{
It assigns the default path to Praat or the user assigned path (when
    \code{default.path = FALSE} (the default).
}

