% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speakr.R
\docType{package}
\name{speakr}
\alias{speakr}
\alias{speakr-package}
\title{speakr: A wrapper for the Phonetic Software Praat}
\description{
It allows running Praat scripts from R and it provides some
wrappers for basic plotting. It also adds support for literate markdown
tangling. The package is designed to bring reproducible phonetic research
into R.
}
