% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praat.R
\name{start_praat}
\alias{start_praat}
\title{Start Praat.}
\usage{
start_praat()
}
\value{
Nothing. Used for its side effects.
}
\description{
It opens the Praat GUI.
}
\examples{
\dontrun{
# Open Praat GUI.
start_praat()
}

}
