% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{use_praat_plot_script}
\alias{use_praat_plot_script}
\title{Use Praat plotting script}
\usage{
use_praat_plot_script(file, ...)
}
\arguments{
\item{file}{Path including file name to which the script is copied.}

\item{...}{Further arguments passed to \code{\link[base]{file.copy}}.}
}
\value{
Nothing. It is used for its side effects.
}
\description{
This allows the user to save a copy of the Praat plotting script on disk so
that they can make changes and customise the plot.
}
