% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{get_praat_path}
\alias{get_praat_path}
\title{Print path to 'Praat'.}
\usage{
get_praat_path()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Now speakr uses options() to set the path to 'Praat'.
}
\keyword{internal}
