% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmt.R
\name{lmt}
\alias{lmt}
\title{Literate markdown tangler.}
\usage{
lmt(file, cache = FALSE)
}
\arguments{
\item{file}{The name of the file to tangle.}

\item{cache}{If set to \code{TRUE}, when the command is run, tangling is not
performed (the default is \code{FALSE}).}
}
\description{
It tangles the code from a literate markdown file. If \code{cache} is set to
    \code{TRUE}, the function does nothing (useful if tangling is not needed
    every time the file is knitted.)
}
