% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praatPath.R
\name{praatPath}
\alias{praatPath}
\title{Get path to Praat.}
\usage{
praatPath()
}
\value{
The path to Praat.
}
\description{
The function returns the path to the Praat executable depeding on the
    system.
}

