% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{set_praat_path}
\alias{set_praat_path}
\title{Set path to 'Praat'.}
\usage{
set_praat_path(path, default.path = FALSE)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Now speakr uses options() to set the path to 'Praat'.
}
\keyword{internal}
