% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{set_praat_path}
\alias{set_praat_path}
\title{Set path to 'Praat'.}
\usage{
set_praat_path(path, default.path = FALSE)
}
\description{
`r lifecycle::badge("deprecated")`

Now speakr uses options() to set the path to 'Praat'.
}
\keyword{internal}
