% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPraatPath.R
\name{setPraatPath}
\alias{setPraatPath}
\title{Set path to Praat}
\usage{
setPraatPath(path, default.path = FALSE)
}
\arguments{
\item{path}{The path to the Praat executable.}

\item{default.path}{If \code{TRUE}, the path is set to what \code{praatPath}
returns.}
}
\description{
It assigns the default path to Praat or the user assigned path (when
    \code{default.path = FALSE}, which is the default).
}
