% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speakr-package.R
\docType{package}
\name{speakr-package}
\alias{speakr}
\alias{speakr-package}
\title{speakr: A Wrapper for the Phonetic Software 'Praat'}
\description{
It allows running 'Praat' scripts from R and it provides some wrappers for basic plotting. It also adds support for literate markdown tangling. The package is designed to bring reproducible phonetic research into R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stefanocoretta/speakr}
  \item Report bugs at \url{https://github.com/stefanocoretta/speakr/issues}
}

}
\author{
\strong{Maintainer}: Stefano Coretta \email{stefano.coretta@gmail.com} (\href{https://orcid.org/0000-0001-9627-5532}{ORCID})

}
\keyword{internal}
